harden runtime ownership and recovery

+++ tests/test_leases.sh
@@ -10,12 +10,14 @@ PMSET_STATE_DIR="$TEST_ROOT/pmset"
 PMSET_LOG="$TEST_ROOT/pmset.log"
 OSASCRIPT_LOG="$TEST_ROOT/osascript.log"
 TEST_HOME="$TEST_ROOT/home"
+TEST_SHELL_PID="$BASHPID"
 mkdir -p "$STUB_BIN" "$STATE_ROOT" "$PMSET_STATE_DIR" "$TEST_HOME/.config/awake"
 
-cleanup() {
+cleanup_test_root() {
+    [ "$BASHPID" = "$TEST_SHELL_PID" ] || return 0
     rm -rf "$TEST_ROOT"
 }
-trap cleanup EXIT
+trap cleanup_test_root EXIT
 
 cat > "$STUB_BIN/sudo" <<'EOF'
 #!/bin/bash
@@ -166,6 +168,7 @@ setup_state() {
     STATE_FILE="$dir/awake-state"
     LAST_ACTIVE_FILE="$dir/awake-last-active"
     CAFFEINE_PID_FILE="$dir/awake-caffeinate.pid"
+    CAFFEINE_PID_START_FILE="$dir/awake-caffeinate-pid-start"
     FOR_PID_FILE="$dir/awake-for.pid"
     FOR_END_FILE="$dir/awake-for-end"
     FOR_TOKEN_FILE="$dir/awake-for-token"
